requirements-dev.txt
langchain
langchain-community
pypdf
sentence-transformers
faiss-cpu
fastapi
uvicorn
python-dotenv
langchain-google-genai